Add the -setActionIsDiscardable: method, as in NSUndoManager in OS X Lion
Invoking method  -setActionIsDiscardable: allows your user to
temporarily change old versions in Viewing Mode in OS X Lion. Because
your user is only viewing the document, NSDocument prevents the user
from changing the model, because that would be changing history. But
sometimes, your user may need to make a change to the view which has, as
a side effect, changing of model data.  That is, sometimes you have some
overlap between your 'model' and your 'view'.  By marking such a change
as 'discardable', the change is allowed. For a "different" explanation
of this, see comments in NSUndoManager.h in the Mac OS 10.7 SDK.