Stop posting NSUndoManagerDidCloseUndoGroupNotification for empty groups, compiles in 10.6
1.  NSUndoManagerDidCloseUndoGroupNotification is now posted only if
there is a current non-empty undo group, mimicking the behavior of
NSUndoManagerWillCloseUndoGroupNotification.  (Note that NSUndoManager
posts both notifications even if the current group is nil or empty; but
we feel that NSUndoManager is wrong in this regard.)    2.  Added
availability macro conditions so that GCUndoManager can now be compiled
using SDKs prior to Mac OS 10.7.    3.  In header, added comments from
Graham's email explaining how to use -setActionName:.